refactor: 도메인과 혼동되는 DTO들의 네이밍 수정
