+++ app/src/main/java/woowacourse/shopping/data/network/cart/dto/CartItemAddRequest.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.data.network.cart.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CartItemAddRequest(
+    @SerialName("productId")
+    val productId: Long,
+    @SerialName("quantity")
+    val quantity: Int,
+)
