+++ app/src/main/java/woowacourse/shopping/data/network/cart/dto/CartItemsResponse.kt
@@ -4,9 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CartItemInsertDto(
-    @SerialName("productId")
-    val productId: Long,
-    @SerialName("quantity")
-    val quantity: Int,
+data class CartItemsResponse(
+    @SerialName("content")
+    val content: List<CartItemDto>,
 )
