+++ app/src/main/java/woowacourse/shopping/data/network/cart/dto/QuantityDto.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Quantity(
+data class QuantityDto(
     @SerialName("quantity")
     val quantity: Int,
 )
