+++ app/src/main/java/woowacourse/shopping/data/network/product/dto/ProductDto.kt
@@ -1,15 +1,22 @@
 package woowacourse.shopping.data.network.product.dto
 
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import woowacourse.shopping.domain.Money
 import woowacourse.shopping.domain.Product
 
 @Serializable
-class ProductDto(
+data class ProductDto(
+    @SerialName("category")
+    val category: String,
+    @SerialName("id")
     val id: Int,
-    val name: String,
-    val price: Int,
+    @SerialName("imageUrl")
     val imageUrl: String,
+    @SerialName("name")
+    val name: String,
+    @SerialName("price")
+    val price: Int,
 ) {
     fun toDomain(): Product {
         return Product(
@@ -17,6 +24,7 @@ class ProductDto(
             price = Money(this.price),
             imageUrl = this.imageUrl,
             id = this.id.toString(),
+            category = this.category,
         )
     }
 }
